Re-doing the implementation of CSS Grid. I'd spotted some places where the content was overflowing out of the max-width I'd set of 660px. It's now much simpler code. I was able to remove the media query.
The gutter down the left and right margins on small screen is still 0.3fr but the central column is now using minmax to define a min-width of auto (i.e. it will fill whatever's left once the 0.3frs have been taken from the sides, and the 1 rem grid gap each side is also taken into account) and will grow to a max of 30 rems.

Added some heavy borders to improve the visual separation on a few sections.

Also refactored the html so that there's a consistent structure across all pages now

body
	|_ header
	|_ main (which includes articles on post pages)
	|_ nav
	|_ footer